day 9

+++ src/main/java/be/selske/aoc2019/days/Day5.java
@@ -26,7 +26,7 @@ public class Day5 extends AocDay {
         return runProgram(5, parseInput(input));
     }
 
-    private static String runProgram(int inputValue, int[] inputs) {
+    private static String runProgram(int inputValue, long[] inputs) {
         IntComputer intComputer = new IntComputer(Arrays.copyOf(inputs, inputs.length));
 
         intComputer.run(inputValue);
@@ -34,9 +34,9 @@ public class Day5 extends AocDay {
         return intComputer.getOutput() + "";
     }
 
-    private static int[] parseInput(Stream<String> input) {
+    private static long[] parseInput(Stream<String> input) {
         return Stream.of(input.findFirst().orElseThrow().split(","))
-                .mapToInt(Integer::valueOf)
+                .mapToLong(Long::valueOf)
                 .toArray();
     }
